Made the single test container run across multiple files
